Squashed commit of the following:
commit b731b16162f47b1516c90ade4d8e78b2583ba252
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Mon Oct 4 15:49:16 2021 +0200

    Fixed warnings.

commit 7c15916740aee6879fe8f692a19737151fcc3d00
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Mon Oct 4 15:45:34 2021 +0200

    Fixed adjusted matrix for stat traceback.

commit 9446695d2879f3aed0d95710292d28a214b14fe8
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Mon Oct 4 14:50:54 2021 +0200

    Bin dptargets by algorithm.

commit 1b20accf6d6c5a5da8690aa1e0375e61e0e792b6
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Thu Sep 30 17:56:54 2021 +0200

    Added test fields.

commit 8af9bdfb4370ae8962d57b82f2403a2f663c19b3
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Thu Sep 30 10:14:16 2021 +0200

    Fixed linear reading for blast db.

commit 20c01c16d3365ccda22d19ffd6c3363635235a2a
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Sep 29 16:18:25 2021 +0200

    Fixed --self option for blast dbs.

commit fda2cf73c9a8ddf708ed38988a9620af2a179215
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Sep 29 15:31:24 2021 +0200

    Fixed alphabet conversion.

commit 4ac0132f0dbb9cd02c6e897b50203e8d13fb82db
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Sep 29 15:13:57 2021 +0200

    Disable comp based stats in self alignment if needed.

commit f6f5aee1879d26385b2c55bcc208e271e6ac052d
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Sep 29 14:10:25 2021 +0200

    Fixed hspvalues for pident_global.

commit a950e84dcce3946b93d3f99d0cd8668ce4c66e38
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Sep 29 14:04:05 2021 +0200

    Fixed compiler errors.

commit f375ad290a54a172e183d111dd0640366e12d8e7
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Sep 29 13:55:34 2021 +0200

    Fixed compiler error.

commit 38c167b9ec14a5d69e3c8245ccfb14935463b1d6
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Sep 29 13:23:05 2021 +0200

    Added pident_global output field.

commit c32ae27594c176b1edd4d73c2a814e38d75c8917
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Sep 29 10:18:05 2021 +0200

    Added global normalized bitscore field.

commit 0b73c17c56a8746b236938faa66f21deb482a000
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Tue Sep 28 16:18:39 2021 +0200

    Fixed compiler errors.

+++ src/align/global_ranking/global_ranking.h
@@ -58,7 +58,7 @@ struct Hit {
 		context((uint8_t)context)
 	{}
 	Hit(ptrdiff_t target_id):
-		oid(target_id),
+		oid((uint32_t)target_id),
 		score()
 	{}
 	uint32_t oid;
