+++ src/cluster/multi_step_cluster.h
@@ -51,10 +51,10 @@ class MultiStep : public ClusteringAlgorithm {
 private:
 	BitVector rep_bitset(const vector<int> &centroid, const BitVector *superset = nullptr);
 	vector<int> cluster(std::shared_ptr<SequenceFile>& db, const std::shared_ptr<BitVector>& filter);
-	void save_edges_external(std::vector<TempFile*> &all_edges,std::vector<TempFile*> &sorted_edges, const std::unordered_map <uint32_t, NodEdgSet>& comp, const vector<uint32_t>& s_index);
+	void save_edges_external(std::vector<TempFile*> &all_edges,std::vector<TempFile*> &sorted_edges, const std::unordered_map <int64_t, NodEdgSet>& comp, const vector<uint32_t>& s_index);
 	vector<int> cluster_sets(const size_t nb_size, vector<TempFile*> &sorted_edges);
-	std::unordered_map<uint32_t, NodEdgSet> find_connected_components(const std::vector<std::unordered_set<uint32_t>> &connected, vector<uint32_t>& EdgSet, const vector<size_t>& nedges);
-	vector<TempFile*> mapping_comp_set(std::unordered_map<uint32_t, NodEdgSet>& comp);
+	std::unordered_map<int64_t, NodEdgSet> find_connected_components(const std::vector<std::unordered_set<int64_t>> &connected, vector<uint32_t>& EdgSet, const vector<size_t>& nedges);
+	vector<TempFile*> mapping_comp_set(std::unordered_map<int64_t, NodEdgSet>& comp);
 	void steps(BitVector& current_reps, BitVector& previous_reps, vector<int>& current_centroids, vector<int>& previous_centroids, int count);
 
 public:
@@ -75,7 +75,7 @@ struct Neighbors : public vector<vector<int>>, public Consumer {
 	vector<size_t> number_edges;
 	vector<TempFile*> tempfiles;
 	size_t size;	
-	LazyDisjointIntegralSet<uint32_t> dSet;
+	LazyDisjointIntegralSet<int64_t> dSet;
 
 
 	virtual void consume(const char* ptr, size_t n) override {
