+++ src/data/reference.cpp
@@ -1,4 +1,4 @@
 #include "reference.h"
 
-unsigned current_ref_block;
+int64_t current_ref_block;
 bool blocked_processing;
